fix(wip): instantiate sequelize once in lambda container
